Split docgen_spec file
